# docs/openapi/panel-api.v1.yaml
openapi: 3.0.3
info:
  title: Lenker Panel API
  version: 0.1.0-draft
  description: Draft OpenAPI contract for the currently implemented Lenker panel-api MVP v0.1 backend surface.
servers:
  - url: http://localhost:8080
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/subscriptions:
    get:
      tags: [Subscriptions]
      summary: List subscriptions
      responses:
        "200":
          description: Subscriptions list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionListEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Subscriptions]
      summary: Create a subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriptionRequest"
      responses:
        "201":
          description: Subscription created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/subscriptions/{id}:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      tags: [Subscriptions]
      summary: Get a subscription
      responses:
        "200":
          description: Subscription details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      tags: [Subscriptions]
      summary: Update a subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSubscriptionRequest"
      responses:
        "200":
          description: Subscription updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/subscriptions/{id}/access:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      tags: [Subscriptions]
      summary: Get provider-side subscription access export
      responses:
        "200":
          description: Subscription access export for the single MVP VLESS Reality path.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionAccessEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Subscription access export is unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/subscriptions/{id}/access-token:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      tags: [Subscriptions]
      summary: Get subscription access token lifecycle status
      description: Returns provider-side lifecycle metadata for the current subscription access token without returning plaintext token material.
      responses:
        "200":
          description: Subscription access token lifecycle status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionAccessTokenStatusEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Subscriptions]
      summary: Issue a subscription access token
      description: Issues the one active subscription access token for this subscription. Any previous active token for the subscription is revoked before the new plaintext token is returned once.
      responses:
        "201":
          description: Plaintext subscription access token created. The database stores only a token hash.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionAccessTokenEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Subscription access token cannot be created for this subscription state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Subscriptions]
      summary: Revoke the active subscription access token
      description: Revokes the current active subscription access token without returning token material. The operation is safe to repeat for an existing subscription.
      responses:
        "200":
          description: Active subscription access token revoked or already absent.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionAccessTokenStatusEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/subscriptions/{id}/access-token/rotate:
    parameters:
      - $ref: "#/components/parameters/ID"
    post:
      tags: [Subscriptions]
      summary: Rotate the active subscription access token
      description: Revokes the previous active token and returns a new plaintext token once.
      responses:
        "201":
          description: Plaintext replacement subscription access token created. The database stores only a token hash.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionAccessTokenEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Subscription access token cannot be rotated for this subscription state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/subscriptions/{id}/renew:
    parameters:
      - $ref: "#/components/parameters/ID"
    post:
      tags: [Subscriptions]
      summary: Manually renew a subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenewSubscriptionRequest"
      responses:
        "200":
          description: Subscription renewed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/client/subscription-access:
    get:
      tags: [Subscriptions]
      summary: Read subscription access with a subscription access token
      description: Consumer-facing read boundary for the current single MVP VLESS Reality path. Uses Authorization Bearer subscription access tokens, not admin sessions.
      responses:
        "200":
          description: Redacted subscription access export.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientSubscriptionAccessEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Subscription access export is unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: session-token
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource identifier.
  responses:
    BadRequest:
      description: Invalid request body or validation failure.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: Missing or invalid admin session token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: Resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    InternalError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [data]
      properties:
        data:
          description: Successful response payload.
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/APIError"
    APIError:
      type: object
      required: [code, message]
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
    ErrorCode:
      type: string
      enum:
        - bad_request
        - unauthorized
        - not_found
        - validation_error
        - storage_error
        - internal_error
        - invalid_credentials
        - inactive_admin
        - invalid_bootstrap_token
        - expired_bootstrap_token
        - bootstrap_token_used
...
    Subscription:
      type: object
      required:
        - id
        - user_id
        - plan_id
        - status
        - starts_at
        - expires_at
        - traffic_used_bytes
        - device_limit
      properties:
        id:
          type: string
        user_id:
          type: string
        plan_id:
          type: string
        status:
          type: string
          enum: [active, expired, suspended]
        starts_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
        traffic_limit_bytes:
          type: integer
          format: int64
          minimum: 1
          nullable: true
        traffic_used_bytes:
          type: integer
          format: int64
          minimum: 0
        device_limit:
          type: integer
          minimum: 1
        preferred_region:
          type: string
          nullable: true
    SubscriptionEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/Subscription"
    SubscriptionListEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/Subscription"
    SubscriptionAccess:
      type: object
      required:
        - export_kind
        - subscription_id
        - user_id
        - user_label
        - plan_id
        - plan_name
        - status
        - protocol
        - protocol_path
        - node
        - endpoint
        - client
        - display_name
        - uri
      properties:
        export_kind:
          type: string
          enum: [subscription_access.v1alpha1]
        subscription_id:
          type: string
        user_id:
          type: string
        user_label:
          type: string
        plan_id:
          type: string
        plan_name:
          type: string
        status:
          type: string
          enum: [active]
        protocol:
          type: string
          enum: [vless-reality-xtls-vision]
        protocol_path:
          type: string
          enum: [vless-reality-xtls-vision]
        node:
          type: object
          required: [id, name, region, country_code, hostname, status, drain_state, active_revision]
          properties:
            id:
              type: string
            name:
              type: string
            region:
              type: string
            country_code:
              type: string
            hostname:
              type: string
            status:
              type: string
              enum: [active]
            drain_state:
              type: string
              enum: [active]
            active_revision:
              type: integer
        endpoint:
          type: object
          required: [address, port, network, security, sni, public_key, short_id, fingerprint, spider_x]
          properties:
            address:
              type: string
            port:
              type: integer
              minimum: 1
            network:
              type: string
              enum: [tcp]
            security:
              type: string
              enum: [reality]
            sni:
              type: string
            public_key:
              type: string
            short_id:
              type: string
            fingerprint:
              type: string
            spider_x:
              type: string
        client:
          type: object
          required: [id, email, flow, level, plan_id]
          properties:
            id:
              type: string
            email:
              type: string
            flow:
              type: string
              enum: [xtls-rprx-vision]
            level:
              type: integer
            plan_id:
              type: string
        display_name:
          type: string
        uri:
          type: string
    SubscriptionAccessEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/SubscriptionAccess"
    SubscriptionAccessToken:
      type: object
      required: [subscription_id, access_token, expires_at, created_at]
      properties:
        subscription_id:
          type: string
        access_token:
          type: string
          description: Plaintext token returned only by the create-token response.
        expires_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    SubscriptionAccessTokenEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/SubscriptionAccessToken"
    SubscriptionAccessTokenStatus:
      type: object
      required: [subscription_id, status, issued, generation]
      properties:
        subscription_id:
          type: string
        status:
          type: string
          enum: [never_issued, active, revoked]
        issued:
          type: boolean
        issued_at:
          type: string
          format: date-time
          nullable: true
        revoked_at:
          type: string
          format: date-time
          nullable: true
        generation:
          type: integer
    SubscriptionAccessTokenStatusEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/SubscriptionAccessTokenStatus"
    ClientSubscriptionAccess:
      type: object
      required:
        - export_kind
        - subscription_id
        - status
        - protocol
        - protocol_path
        - plan_name
        - node
        - endpoint
        - client
        - display_name
        - uri
      properties:
        export_kind:
          type: string
          enum: [subscription_access.v1alpha1]
        subscription_id:
          type: string
        status:
          type: string
          enum: [active]
        protocol:
          type: string
          enum: [vless-reality-xtls-vision]
        protocol_path:
          type: string
          enum: [vless-reality-xtls-vision]
        plan_name:
          type: string
        node:
          $ref: "#/components/schemas/SubscriptionAccess/properties/node"
        endpoint:
          $ref: "#/components/schemas/SubscriptionAccess/properties/endpoint"
        client:
          type: object
          required: [id, email, flow, level]
          properties:
            id:
              type: string
            email:
              type: string
            flow:
              type: string
              enum: [xtls-rprx-vision]
            level:
              type: integer
        display_name:
          type: string
        uri:
          type: string
    ClientSubscriptionAccessEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/ClientSubscriptionAccess"
    CreateSubscriptionRequest:
      type: object
      required: [user_id, plan_id]
      properties:
        user_id:
          type: string
        plan_id:
          type: string
        preferred_region:
          type: string
          nullable: true
    UpdateSubscriptionRequest:
      type: object
      properties:
        status:
          type: string
          enum: [active, expired, suspended]
        traffic_limit_bytes:
          type: integer
          format: int64
          minimum: 1
          nullable: true
        clear_traffic_limit:
          type: boolean
        device_limit:
          type: integer
          minimum: 1
        preferred_region:
          type: string
          nullable: true
        clear_preferred_region:
          type: boolean
    RenewSubscriptionRequest:
      type: object
      required: [extend_days]
      properties:
        extend_days:
          type: integer
          minimum: 1
...